+++ src/Model/DAO/AccountDAO.java
@@ -1,0 +1,51 @@
+
+package Model.DAO;
+
+import java.util.ArrayList;
+
+import Model.Entity.Account;
+
+public class AccountDAO implements iDAO<Account> {
+    private ArrayList<Account> accList;
+
+    public AccountDAO() {
+        accList = readFromFile();
+    }
+
+    @Override
+    public ArrayList<Account> getList() {
+        throw new UnsupportedOperationException("Not supported yet."); // To change body of generated methods, choose
+                                                                       // Tools | Templates.
+    }
+
+    @Override
+    public ArrayList<Account> readFromFile() {
+        throw new UnsupportedOperationException("Not supported yet."); // To change body of generated methods, choose
+                                                                       // Tools | Templates.
+    }
+
+    @Override
+    public boolean exist(Account object) {
+        throw new UnsupportedOperationException("Not supported yet."); // To change body of generated methods, choose
+                                                                       // Tools | Templates.
+    }
+
+    @Override
+    public void add(Account object) {
+        throw new UnsupportedOperationException("Not supported yet."); // To change body of generated methods, choose
+                                                                       // Tools | Templates.
+    }
+
+    @Override
+    public void delete(Account object) {
+        throw new UnsupportedOperationException("Not supported yet."); // To change body of generated methods, choose
+                                                                       // Tools | Templates.
+    }
+
+    @Override
+    public void writeToFile() {
+        throw new UnsupportedOperationException("Not supported yet."); // To change body of generated methods, choose
+                                                                       // Tools | Templates.
+    }
+
+}
