+++ src/Model/DAO/DealerDAO.java
@@ -1,4 +1,4 @@
-package Model;
+package Model.DAO;
 
 public class DealerDAO {
