+++ src/Model/Entity/Account.java
@@ -1,4 +1,4 @@
-package Model;
+package Model.Entity;
 
 public class Account {
